ensure notebook runs even without some of the temp Duke modificaitons that don't have a home in the code yet.
